refactor(postgres): lineage and its manager split along their own section rules
lineage.rb was six responsibilities and a thousand lines; each of its
comment-delimited sections is now a module under lineage/ —
provisioning (DDL and the RLS posture), era_store (hecks_eras rows and
their integrity), mint_transaction (the one transaction that makes an
era real), tail_merge, head_compiler (the chained-edge SQL), and
transform_installer (the hecks_tr_* helpers). The outside file keeps
the topology narrative, the naming scheme, and the quoting helpers.
LineageManager likewise: era_resolver, minter, merge_coordinator and
coverage_check under lineage_manager/, with the shared edge-chain walk
and shadow parse staying on the face. Full parity, era dance included,
green before this commit.
